# deno/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "skmtc stack server",
    "version": "1.1.0",
    "description": "The public API of a deployed @skmtc/server bundle: generate code from an OpenAPI/GraphQL schema against the bundled generator stack, plus schema, capability and enrichment introspection. Codegen is a pure function of (bundle, schema, client settings)."
  },
  "paths": {
    "/artifacts": {
      "post": {
        "operationId": "generateArtifacts",
        "summary": "Generate code artifacts from a schema",
        "description": "Parse → transform → render the schema through the bundled generators, returning the generated files, the run manifest, and the attribution sidecars + generation map. The schema arrives inline (`schema`) or by URL (`source` — fetched by the server, echoed back with its resolved URL and content digest); `protocol` is inferred from the document when omitted.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated artifacts + manifest + attribution.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "artifacts",
                    "manifest"
                  ],
                  "properties": {
                    "artifacts": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      },
                      "description": "Generated file contents keyed by output path."
                    },
                    "manifest": {
                      "$ref": "#/components/schemas/Manifest"
                    },
                    "sidecars": {
                      "type": "object",
                      "additionalProperties": {
                        "$ref": "#/components/schemas/Sidecar"
                      },
                      "description": "Per-file attribution sidecars keyed by output path."
                    },
                    "generationMap": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/GenerationMapEntry"
                      },
                      "description": "Per-Definition generation-map index (file → schema origin)."
                    },
                    "source": {
                      "$ref": "#/components/schemas/ResolvedSource"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed request — invalid JSON body or failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable — the `source` could not be fetched, or the document could not be read as a schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/subjects": {
      "post": {
        "operationId": "supportedSubjects",
        "summary": "Which subjects each generator supports",
        "description": "Parse + each generator’s `isSupported` only — no transform, no render.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Per-generator supported subjects + parse issues.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "subjects",
                    "parseIssues"
                  ],
                  "properties": {
                    "subjects": {
                      "type": "object",
                      "additionalProperties": {
                        "$ref": "#/components/schemas/GeneratorSupport"
                      },
                      "description": "Supported subjects keyed by generator id."
                    },
                    "parseIssues": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ParseIssue"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed request — invalid JSON body or failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable — the `source` could not be fetched, or the document could not be read as a schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/enrichment-defaults": {
      "post": {
        "operationId": "enrichmentDefaults",
        "summary": "Seed enrichment values derived from a schema",
        "description": "The default enrichment values each generator derives from the schema (the \"Generate fields from schema\" payload) — Parse + `toEnrichmentDefaults` over supported subjects, no transform, no render.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Seed enrichment values + parse issues.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "enrichmentDefaults",
                    "parseIssues"
                  ],
                  "properties": {
                    "enrichmentDefaults": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Seed enrichment values, mirroring the `client.json#settings.enrichments` subtree (subject scope): `[generatorId][path][method][variant]` for operations, `[generatorId][refName][variant]` for models."
                    },
                    "parseIssues": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ParseIssue"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed request — invalid JSON body or failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable — the `source` could not be fetched, or the document could not be read as a schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/descriptors": {
      "post": {
        "operationId": "enrichmentDescriptors",
        "summary": "Form-renderable enrichment descriptors",
        "description": "The renderable descriptor for each generator’s enrichment schema — a pure function of the bundled generators (no schema, no parse), safe to cache per bundle. POST with no body.",
        "responses": {
          "200": {
            "description": "One descriptor per configured generator.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "descriptors"
                  ],
                  "properties": {
                    "descriptors": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EnrichmentDescriptor"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed request — invalid JSON body or failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable — the `source` could not be fetched, or the document could not be read as a schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/validate": {
      "post": {
        "operationId": "validateEnrichments",
        "summary": "Validate an enrichment config against the stack",
        "description": "The authoritative verdict on whether the posted `clientSettings.enrichments` satisfy each generator’s enrichment schema. Documentless — no schema needed.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validation issues (empty when the config is valid).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "issues"
                  ],
                  "properties": {
                    "issues": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EnrichmentValidationIssue"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed request — invalid JSON body or failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable — the `source` could not be fetched, or the document could not be read as a schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/to-v3-json": {
      "post": {
        "operationId": "toV3Json",
        "summary": "Normalize a schema document to OpenAPI 3.0",
        "description": "Convert a Swagger 2.0 document to OpenAPI 3.0; 3.0 / 3.1 documents pass through unchanged. The same normalization the `oas` generate path runs.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToV3JsonRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The normalized OpenAPI 3.0 document.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "schema"
                  ],
                  "properties": {
                    "schema": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "The input document normalized to an OpenAPI 3.0 document."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed request — invalid JSON body or failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable — the `source` could not be fetched, or the document could not be read as a schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ArtifactsRequest": {
        "type": "object",
        "properties": {
          "protocol": {
            "enum": [
              "oas",
              "gql"
            ],
            "description": "Schema protocol. Optional — inferred from the document content when omitted (a JSON/YAML document with an `openapi` or `swagger` key is `oas`; anything else is treated as GraphQL SDL)."
          },
          "schema": {
            "type": "string",
            "description": "The schema document itself, as a string."
          },
          "source": {
            "type": "string",
            "format": "uri",
            "description": "URL of the schema document. The server fetches it (following redirects) and echoes the final URL + content digest in the response."
          },
          "clientSettings": {
            "type": "object",
            "properties": {
              "basePath": {
                "type": "string"
              },
              "schemaSource": {
                "type": "string"
              },
              "packages": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "rootPath": {
                      "type": "string"
                    },
                    "moduleName": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "rootPath"
                  ]
                }
              },
              "enrichments": {
                "type": "object",
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "additionalProperties": {}
                      }
                    },
                    {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "additionalProperties": {}
                      }
                    }
                  ]
                }
              },
              "include": {
                "type": "array",
                "items": {
                  "anyOf": [
                    {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "additionalProperties": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    },
                    {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "additionalProperties": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    {
                      "type": "string"
                    }
                  ]
                }
              },
              "skip": {
                "type": "array",
                "items": {
                  "anyOf": [
                    {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "additionalProperties": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    },
                    {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "additionalProperties": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    {
                      "type": "string"
                    }
                  ]
                }
              },
              "anchors": {
                "type": "object",
                "properties": {
                  "enabled": {
                    "type": "boolean"
                  },
                  "out": {
                    "type": "string"
                  }
                },
                "required": [
                  "enabled"
                ]
              },
              "inputDirs": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "formatter": {
                "type": "string"
              },
              "generatedSuffix": {
                "type": "string"
              },
              "ejected": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": []
          },
          "schemaSrc": {
            "type": "string"
          }
        },
        "required": []
      },
...
      "Sidecar": {
        "type": "object",
        "properties": {
          "v": {
            "const": 2
          },
          "f": {
            "type": "string"
          },
          "src": {
            "type": "string"
          },
          "parser": {
            "type": "string"
          },
          "R": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "host": {
                  "type": "string"
                },
                "type": {
                  "anyOf": [
                    {
                      "const": "jsr"
                    },
                    {
                      "const": "jsr-private"
                    }
                  ]
                }
              },
              "required": [
                "host",
                "type"
              ]
            }
          },
          "G": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "version": {
                  "type": "string"
                },
                "r": {
                  "type": "number"
                }
              },
              "required": [
                "name",
                "version",
                "r"
              ]
            }
          },
          "S": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "V": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "L": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "P": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "A": {
            "type": "array",
            "items": {
              "type": "array",
              "items": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                },
                {
                  "type": "number"
                },
                {
                  "type": "number"
                },
                {
                  "type": "number"
                },
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "minItems": 7
            }
          },
          "N": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "An": {
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "v",
          "f",
          "src",
          "parser",
          "R",
          "G",
          "S",
          "V",
          "L",
          "P",
          "A"
        ]
      },
      "GenerationMapEntry": {
        "type": "object",
        "properties": {
          "f": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "g": {
            "type": "string"
          },
          "s": {
            "type": "string"
          },
          "v": {
            "type": "string"
          }
        },
        "required": [
          "f",
          "name",
          "g",
          "s",
          "v"
        ]
      },
...
    }
  }
}